The rendered decorator is now passed through the ContentProcessor to, to further process any tags added by the decorator. This allows decorators to be built with tags and treated as content, rather than having bindings for templating languages (JSP, Velocity...).
Some cool side effects of this:
* No more never-ending list of templating systems to integrate with.
* Works in the same way in web-apps and offline.
* It's just so simple.

Overview of changes:
* Introduced SiteMeshWriteTagRule, which handles <sitemesh:write property='body'/>
definitions in decorators and merges in the content.
* <decorate> renamed to <sitemesh:decorate>.
* SiteMesh Context object now has a getContentToMerge() method which is used by
decorators when they need to get access to the underlying content.
* Content now has a 'processed' property which is the default output (post-processing).
Previously this was implicitly 'body' - just making it more explicit.

Note: This changes introduces a bug and a bunch of test failures around the <title>
tag. These will be resolved in a follow up commit (this change is already very big)
and will not be pushed upstream until fixed.