Externalized tag name in TagRule:
Rather than TagRule having to implement a canHandle(String tagName) method, it is down to the State to determine which TagRule to match up. By externalizing this, it makes the code easier to read (i.e. by looking at a ContentProcessor implementation, it is obvious which tags are handled by the rules) and also allows for the State to make performance improvements (i.e. using a hash lookup, instead of looping through each rule for each tag).